rf: move GenerateSamplingReference into nibabel module
