+++ niworkflows/interfaces/nibabel.py
@@ -284,3 +284,152 @@ class FilledImageLike(SimpleInterface):
             newpath=runtime.cwd,
         )
         return runtime
+
+
+class _GenerateSamplingReferenceInputSpec(BaseInterfaceInputSpec):
+    fixed_image = File(
+        exists=True, mandatory=True, desc="the reference file, defines the FoV"
+    )
+    moving_image = File(exists=True, mandatory=True, desc="the pixel size reference")
+    xform_code = traits.Enum(None, 2, 4, usedefault=True, desc="force xform code")
+    fov_mask = traits.Either(
+        None,
+        File(exists=True),
+        usedefault=True,
+        desc="mask to clip field of view (in fixed_image space)",
+    )
+    keep_native = traits.Bool(
+        True,
+        usedefault=True,
+        desc="calculate a grid with native resolution covering "
+        "the volume extent given by fixed_image, fast forward "
+        "fixed_image otherwise.",
+    )
+
+
+class _GenerateSamplingReferenceOutputSpec(TraitedSpec):
+    out_file = File(exists=True, desc="one file with all inputs flattened")
+
+
+class GenerateSamplingReference(SimpleInterface):
+    """
+    Generates a reference grid for resampling one image keeping original resolution,
+    but moving data to a different space (e.g. MNI).
+
+    If the `fov_mask` optional input is provided, then the abbr:`FoV (field-of-view)`
+    is cropped to a bounding box containing the brain mask plus an offest of two
+    voxels along all dimensions. The `fov_mask` should be to the brain mask calculated
+    from the T1w, and should not contain the brain stem. The mask is resampled into
+    target space, and then the bounding box is calculated. Finally, the FoV is adjusted
+    to that bounding box.
+
+
+    """
+
+    input_spec = _GenerateSamplingReferenceInputSpec
+    output_spec = _GenerateSamplingReferenceOutputSpec
+
+    def _run_interface(self, runtime):
+        if not self.inputs.keep_native:
+            self._results["out_file"] = self.inputs.fixed_image
+            return runtime
+
+        from .. import __version__
+
+        self._results["out_file"] = _gen_reference(
+            self.inputs.fixed_image,
+            self.inputs.moving_image,
+            fov_mask=self.inputs.fov_mask,
+            force_xform_code=self.inputs.xform_code,
+            message="%s (niworkflows v%s)" % (self.__class__.__name__, __version__),
+            newpath=runtime.cwd,
+        )
+        return runtime
+
+
+def _gen_reference(
+    fixed_image,
+    moving_image,
+    fov_mask=None,
+    out_file=None,
+    message=None,
+    force_xform_code=None,
+    newpath=None,
+):
+    """Generate a sampling reference, and makes sure xform matrices/codes are correct."""
+    import nilearn.image as nli
+
+    if out_file is None:
+        out_file = fname_presuffix(
+            fixed_image, suffix="_reference", newpath=newpath
+        )
+
+    # Moving images may not be RAS/LPS (more generally, transverse-longitudinal-axial)
+    reoriented_moving_img = nb.as_closest_canonical(nb.load(moving_image))
+    new_zooms = reoriented_moving_img.header.get_zooms()[:3]
+
+    # Avoid small differences in reported resolution to cause changes to
+    # FOV. See https://github.com/nipreps/fmriprep/issues/512
+    # A positive diagonal affine is RAS, hence the need to reorient above.
+    new_affine = np.diag(np.round(new_zooms, 3))
+
+    resampled = nli.resample_img(
+        fixed_image, target_affine=new_affine, interpolation="nearest"
+    )
+
+    if fov_mask is not None:
+        # If we have a mask, resample again dropping (empty) samples
+        # out of the FoV.
+        fixednii = nb.load(fixed_image)
+        masknii = nb.load(fov_mask)
+
+        if np.all(masknii.shape[:3] != fixednii.shape[:3]):
+            raise RuntimeError("Fixed image and mask do not have the same dimensions.")
+
+        if not np.allclose(masknii.affine, fixednii.affine, atol=1e-5):
+            raise RuntimeError("Fixed image and mask have different affines")
+
+        # Get mask into reference space
+        masknii = nli.resample_img(
+            masknii, target_affine=new_affine, interpolation="nearest"
+        )
+        res_shape = np.array(masknii.shape[:3])
+
+        # Calculate a bounding box for the input mask
+        # with an offset of 2 voxels per face
+        bbox = np.argwhere(np.asanyarray(masknii.dataobj) > 0)
+        new_origin = np.clip(bbox.min(0) - 2, a_min=0, a_max=None)
+        new_end = np.clip(bbox.max(0) + 2, a_min=0, a_max=res_shape - 1)
+
+        # Find new origin, and set into new affine
+        new_affine_4 = resampled.affine.copy()
+        new_affine_4[:3, 3] = new_affine_4[:3, :3].dot(new_origin) + new_affine_4[:3, 3]
+
+        # Calculate new shapes
+        new_shape = new_end - new_origin + 1
+        resampled = nli.resample_img(
+            fixed_image,
+            target_affine=new_affine_4,
+            target_shape=new_shape.tolist(),
+            interpolation="nearest",
+        )
+
+    xform = resampled.affine  # nibabel will pick the best affine
+    _, qform_code = resampled.header.get_qform(coded=True)
+    _, sform_code = resampled.header.get_sform(coded=True)
+
+    xform_code = sform_code if sform_code > 0 else qform_code
+    if xform_code == 1:
+        xform_code = 2
+
+    if force_xform_code is not None:
+        xform_code = force_xform_code
+
+    # Keep 0, 2, 3, 4 unchanged
+    resampled.header.set_qform(xform, int(xform_code))
+    resampled.header.set_sform(xform, int(xform_code))
+    resampled.header["descrip"] = "reference image generated by %s." % (
+        message or "(unknown software)"
+    )
+    resampled.to_filename(out_file)
+    return out_file
